// Frontend-App/final-project/src/app/models/laptop.model.ts
export interface Laptop {
  image?: any;
  brand?: string;
  model?: string;
  price: number;
  url?: string;
  fabric_date?: string;
  cpu?: string;
  cpu_core?: number;
  cpu_speed?: number;
  ram?: string;
  ram_type?: string;
  ram_size?: number;
  ram_expand?: number;
  ram_speed?: number;
  rom?: string;
  rom_type?: string;
  rom_capacity?: string;
  graphics?: string;
  graphics_brand?: string;
  graphics_model?: string;
  resolution?: string;
  battery?: number;
  os?: string;
  weight?: number;
  _id?: any;
}